requirements.txt
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.3
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
